requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
flasgger==0.9.5
Flask==2.2.5
flask-cors==5.0.1
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
flask-openapi3==2.5.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
MarkupSafe==3.0.2
mistune==3.1.2
multidict==6.1.0
openai==0.28.0
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
psycopg2-binary==2.9.9
pydantic==1.10.8
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==7.4.3
python-dotenv==1.0.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
rpds-py==0.23.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==2.2.3
yarl==1.18.3
zipp==3.21.0
